Split appsync.py into 2 files
Modify AIOHTTPTransport to accept an AppSync auth class
