+++ docs/modules/gql.rst
@@ -21,7 +21,8 @@ Sub-Packages
    client
    transport
    transport_aiohttp
-   transport_appsync
+   transport_appsync_auth
+   transport_appsync_websockets
    transport_exceptions
    transport_phoenix_channel_websockets
    transport_requests
