--- docs/modules/transport_appsync.rst
@@ -1,7 +1,0 @@
-gql.transport.appsync
-=====================
-
-.. currentmodule:: gql.transport.appsync
-
-.. automodule:: gql.transport.appsync
-    :member-order: bysource
