+++ docs/modules/transport_appsync_auth.rst
@@ -1,0 +1,7 @@
+gql.transport.appsync_auth
+==========================
+
+.. currentmodule:: gql.transport.appsync_auth
+
+.. automodule:: gql.transport.appsync_auth
+    :member-order: bysource
