+++ docs/modules/transport_appsync_websockets.rst
@@ -1,0 +1,7 @@
+gql.transport.appsync_websockets
+================================
+
+.. currentmodule:: gql.transport.appsync_websockets
+
+.. automodule:: gql.transport.appsync_websockets
+    :member-order: bysource
